# api/openapi.yaml
openapi: 3.0.3
info:
  title: LabLDAP Management API
  version: 1.0.0
  description: |
    Versioned control-plane REST contract for LabLDAP.

    OpenAPI 3.0.3 is the documented compatible subset of 3.1 (OD-009).
    Payload shapes follow design §3.8 (`x-labldap-contract: proposed`
    stand-in dated 2026-08-13). Architecture-named (binding) routes are
    unmarked. Every other operation is marked `x-labldap-contract: proposed`.

    There is no `PATCH /api/v1/groups/{id}` in v1. Membership mutations
    are the group update path.

    Examples use lab placeholders only. They are not usable credentials.
  license:
    name: UNLICENSED
    url: https://github.com/hilather/go-lab-ldap-mcp
...
servers:
  - url: /
    description: Same-origin control listener
...
paths:
...
  /api/v1/capabilities:
    get:
      operationId: getCapabilities
      tags: [System]
      summary: Measured engine capabilities
      x-labldap-contract: proposed
      x-required-scope: directory:read
      security:
        - bearerAuth: [directory:read]
        - cookieAuth: []
      responses:
        "200":
          description: Capabilities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capabilities"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/baseline:
    get:
      operationId: getBaseline
      tags: [System]
      summary: Compiled versus applied baseline
      x-labldap-contract: proposed
      x-required-scope: directory:read
      security:
        - bearerAuth: [directory:read]
        - cookieAuth: []
      responses:
        "200":
          description: Baseline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Baseline"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/session:
    post:
      operationId: createSession
      tags: [Session]
      summary: Exchange a static token for a browser session
      description: |
        Binding `POST /api/v1/session`. Body is `{ "token": "<secret>" }` only.
        Sets cookie `labldap_session` and returns `{ "csrfToken": "..." }`.
        Never returns the raw token. Login rotates any existing session.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionCreate"
            example:
              token: lab-example-static-token
      responses:
        "200":
          description: Session created
          headers:
            Set-Cookie:
              schema:
                type: string
              description: labldap_session HttpOnly SameSite=Strict Path=/
            Cache-Control:
              schema:
                type: string
                example: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionCreated"
              example:
                csrfToken: lab-example-csrf-not-a-secret-from-prod
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    get:
      operationId: getSession
      tags: [Session]
      summary: Current session
      x-labldap-contract: proposed
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        "200":
          description: Non-secret session view
          headers:
            Cache-Control:
              schema:
                type: string
                example: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionView"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteSession
      tags: [Session]
      summary: Logout
      x-labldap-contract: proposed
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      responses:
        "204":
          description: Session invalidated and cookie cleared
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/users:
    get:
      operationId: listUsers
      tags: [Users]
      summary: List users
      x-labldap-contract: proposed
      x-required-scope: directory:read
      security:
        - bearerAuth: [directory:read]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Cursor"
        - name: q
          in: query
          schema:
            type: string
          description: Optional substring on id or uid
      responses:
        "200":
          description: User page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createUser
      tags: [Users]
      summary: Create a user
      description: Binding `POST /api/v1/users`. Returns 201, Location, and ETag.
      x-required-scope: directory:write
      security:
        - bearerAuth: [directory:write]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSpec"
            example:
              id: alice
              uid: alice
              enabled: true
              password: lab-example-password
              attributes: {}
      responses:
        "201":
          description: Created
          headers:
            Location:
              schema:
                type: string
                example: /api/v1/users/alice
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/users/{id}:
    parameters:
      - $ref: "#/components/parameters/UserID"
    get:
      operationId: getUser
      tags: [Users]
      summary: Get a user
      x-labldap-contract: proposed
      x-required-scope: directory:read
      security:
        - bearerAuth: [directory:read]
        - cookieAuth: []
      responses:
        "200":
          description: User
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateUser
      tags: [Users]
      summary: Patch a user (not password)
      x-labldap-contract: proposed
      x-required-scope: directory:write
      security:
        - bearerAuth: [directory:write]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPatch"
      responses:
        "200":
          description: Updated
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
    delete:
      operationId: deleteUser
      tags: [Users]
      summary: Delete a user
      x-labldap-contract: proposed
      x-required-scope: directory:write
      security:
        - bearerAuth: [directory:write]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
  /api/v1/users/{id}/password:
    post:
      operationId: setUserPassword
      tags: [Users]
      summary: Set a user password. Optional mustChange forces a reset on next bind.
      x-labldap-contract: proposed
      x-required-scope: directory:password
      security:
        - bearerAuth: [directory:password]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordBody"
            example:
              password: lab-example-password
              revision: "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
      responses:
        "204":
          description: Password updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/users/{id}/enable:
    post:
      operationId: enableUser
      tags: [Users]
      summary: Enable a user
      x-labldap-contract: proposed
      x-required-scope: directory:write
      security:
        - bearerAuth: [directory:write]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200":
          description: Enabled
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
  /api/v1/users/{id}/disable:
    post:
      operationId: disableUser
      tags: [Users]
      summary: Disable a user
      x-labldap-contract: proposed
      x-required-scope: directory:write
      security:
        - bearerAuth: [directory:write]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200":
          description: Disabled
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
  /api/v1/users/{id}/account-state:
    get:
      operationId: getUserAccountState
      tags: [Users]
      summary: Read enable, lock, and must-change state
      x-labldap-contract: proposed
      x-required-scope: directory:read
      security:
        - bearerAuth: [directory:read]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: Account workflow state
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountState"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/users/{id}/expire-password:
    post:
      operationId: expireUserPassword
      tags: [Users]
      summary: Force must-change on next successful bind-test
      x-labldap-contract: proposed
      x-required-scope: directory:password
      security:
        - bearerAuth: [directory:password]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200":
          description: Password marked must-change
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountState"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
  /api/v1/users/{id}/clear-password-expiry:
    post:
      operationId: clearUserPasswordExpiry
      tags: [Users]
      summary: Clear must-change without setting a new password
      x-labldap-contract: proposed
      x-required-scope: directory:password
      security:
        - bearerAuth: [directory:password]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200":
          description: Must-change cleared
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountState"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
  /api/v1/users/{id}/lock:
    post:
      operationId: lockUser
      tags: [Users]
      summary: Administratively lock a user
      x-labldap-contract: proposed
      x-required-scope: directory:write
      security:
        - bearerAuth: [directory:write]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200":
          description: Locked
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountState"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
  /api/v1/users/{id}/unlock:
    post:
      operationId: unlockUser
      tags: [Users]
      summary: Clear lockout stamps
      x-labldap-contract: proposed
      x-required-scope: directory:write
      security:
        - bearerAuth: [directory:write]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "200":
          description: Unlocked
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountState"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
  /api/v1/users/{id}/groups:
    get:
      operationId: listUserGroups
      tags: [Users]
      summary: Groups for a user
      x-labldap-contract: proposed
      x-required-scope: directory:read
      security:
        - bearerAuth: [directory:read]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: Group list envelope
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/groups:
    get:
      operationId: listGroups
      tags: [Groups]
      summary: List groups
      x-labldap-contract: proposed
      x-required-scope: directory:read
      security:
        - bearerAuth: [directory:read]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Cursor"
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Group page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createGroup
      tags: [Groups]
      summary: Create a group
      x-labldap-contract: proposed
      x-required-scope: directory:write
      security:
        - bearerAuth: [directory:write]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupSpec"
            example:
              id: staff
              members:
                - kind: user
                  id: alice
      responses:
        "201":
          description: Created
          headers:
            Location:
              schema:
                type: string
                example: /api/v1/groups/staff
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/groups/{id}:
    parameters:
      - $ref: "#/components/parameters/GroupID"
    get:
      operationId: getGroup
      tags: [Groups]
      summary: Get a group
      x-labldap-contract: proposed
      x-required-scope: directory:read
      security:
        - bearerAuth: [directory:read]
        - cookieAuth: []
      responses:
        "200":
          description: Group
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteGroup
      tags: [Groups]
      summary: Delete a group
      x-labldap-contract: proposed
      x-required-scope: directory:write
      security:
        - bearerAuth: [directory:write]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IfMatch"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
  /api/v1/groups/{id}/members:
    parameters:
      - $ref: "#/components/parameters/GroupID"
    post:
      operationId: addGroupMembers
      tags: [Groups]
      summary: Add members
      x-labldap-contract: proposed
      x-required-scope: directory:write
      security:
        - bearerAuth: [directory:write]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberList"
      responses:
        "200":
          description: Membership summary
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
    delete:
      operationId: removeGroupMembers
      tags: [Groups]
      summary: Remove members
      x-labldap-contract: proposed
      x-required-scope: directory:write
      security:
        - bearerAuth: [directory:write]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberList"
      responses:
        "200":
          description: Membership summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
    put:
      operationId: replaceGroupMembers
      tags: [Groups]
      summary: Replace members
      x-labldap-contract: proposed
      x-required-scope: directory:write
      security:
        - bearerAuth: [directory:write]
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberList"
      responses:
        "200":
          description: Membership summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembershipSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
...
  /api/v1/rootdse:
    get:
      operationId: getRootDSE
      tags: [Schema]
      summary: Root DSE
      x-labldap-contract: proposed
      x-required-scope: schema:read
      security:
        - bearerAuth: [schema:read]
        - cookieAuth: []
      responses:
        "200":
          description: Root DSE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RootDSE"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Static lab token from a secret file (OD-013). Compared in constant
        time. Only the non-secret token ID is exposed after authentication.
        Scopes: directory:read, directory:write, directory:password,
        lab:reset, lab:export, schema:read, audit:read.
    cookieAuth:
      type: apiKey
      in: cookie
      name: labldap_session
      description: |
        Opaque in-memory session cookie. HttpOnly, SameSite=Strict,
        Secure when TLS, Path=/. Cookie-authenticated mutations require
        header X-CSRF-Token and a valid Origin.
...
  responses:
...
    Unauthorized:
      description: Missing, malformed, or invalid credentials
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://labldap.dev/problems/auth
            title: authentication required
            status: 401
            extensions:
              requestId: 00000000000000000000000000000000
    Forbidden:
      description: Missing scope, CSRF, or Origin
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
security:
  - bearerAuth: []
  - cookieAuth: []